% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glmrgame-package.r
\docType{package}
\name{glmrgame-package}
\alias{glmrgame-package}
\title{glmrgame}
\description{
Linear and generalized linear models (and support vector machine)
for 'GPUs'.
}
\references{
Programming with Big Data in R Website: \url{https://pbdr.org/}
}
\keyword{Package}
